Fixing crash when SimpleShipTest.grc and it's variants exit
Function probes have multiple race conditions affecting them one of
which was caused a crash when the the script exits. The probe has been
replaced with an alternative that is not affect by this issue.

There was talk of depricating function probes in git issues related to
them so we should stop using them.